Initial commit
Just a very rough CLI implementation of a tool idea based on Cargo's [1]
`update` subcommand [2]. The tool aims to update transitive dependencies
in npm projects in bulk. I'm not aware of any such tool, and in my
experience the npm CLI does not currently provide a convenient way to do
this (currently, I intentionally licensed this project so that it would
be easy for the npm CLI to adopt this, the idea if not the code).

The implementation is currently very crude: argument parsing is very
simplistic (and definitely not 100% correct), the updating approach is
probably too coarse (in writing and testing I've already seen some
obvious bugs, e.g.: in a project with one dependency (of a given type)
this will cause that dependency type to be moved to the bottom of the
manifest since uninstall the dependency results in the removal of the
dependency type object entirely, in turn adding will cause that object
to appear at the end).

The name is a play on "UPDATE Transitive Dependencies" and verb
conjugation(?) in Dutch, where the "-dt" suffix is sometimes used for
verbs if the subject of the sentence is in the second person. It's a
long shot, but who cares. While we're on the topic, I hereby declare
this _should_ be pronounced the same as "updated".

--
1. https://doc.rust-lang.org/cargo/
2. https://doc.rust-lang.org/cargo/commands/cargo-update.html