Improve CLI argument parsing
This adds a proper CLI arguments parser: meow [1]. This is a pretty
common (at the time of this commit >19M weekly downloads) dependency
for this purpose which is simple to use while effective. The latest
version is used to maximize features, support, and longevity.

With the first project dependency also comes a lockfile and the `.npmrc`
file used to configure the lockfile (the used option requires at least
v8.1.2 of the npm CLI). Git has also been instructed to not show changes
to the machine generated lockfile by default. More details on this
project's approach to dependency management are documented in the new
`docs/dev/dependencies.md` file.

Using a CLI arguments parser library enables:
- Improved `--help` message.
- A `--version` flag pulled from this project's `package.json`.
- Unknown flag detection.
- Use of non-boolean flags.
- Use of repeatable flags.

This addresses a misalignment between how this CLI scoped its behavior
to dependency catagories and how the official npm CLI does. Like the
official npm CLI, this CLI now has a `--omit` flag that takes an enum of
values (validated) to exclude certain dependency categories.

NOTE: As a result the `--prod-only` and `--dev-only` flags have been
      removed.

This commit also includes the first changelog. The changelog is modelled
after the _Keep a Changelog_ project's [2] template. The changes for
the first version simply comprise the features included in the initial
release and the date is just the date of the initial commit.

Stray changes:
- Expand .gitignore to include file system files that should always be
  ignored.
- Make purpose of `console.log()` call explicit.
- Formatting corrections to the `package.json` (removed newline, added
  space after colon).
- Version bump due to user facing changes.
- Explicit SPDX license identifier header to source code.
- Explicit documentation licenses.
- Fix exit codes values and usage.

--
1. https://www.npmjs.com/package/meow
2. https://keepachangelog.com/en/1.0.0/