Fix failure on missing (prod|dev) dependencies
Implement a fix for the bug where a project missing production or
development dependencies ("dependencies" or "devDependencies" key in
`package.json`) results in an unexpected error, terminating the program
early. Fixed by checking if "(dev)Dependencies" are defined, and not
trying to update if that isn't the case.

This is accompanied by the initialization of a test suite for the
project. Following the offical npm CLI, the test runner is tap [1].
The test approach used is snapshot testing, natively supported by
tap [2], where values from previous test runs are stored and compared
against current outputs. For now, stdout and exit codes are tested. In
particular because both of these were incorrect prior to the changes to
`main.js`.

--
1. https://node-tap.org/
2. https://node-tap.org/docs/api/snapshot-testing/