Recognize `npm update`
It turns out this project was for naught! npm already has support for
doing this kind of dependency management through the `npm update`
command. I recall this subcommand existed but thought it didn't work
well. Nevertheless, I gave it a try and it turns out it does everything
I want from it.

This commit just adds a quick note to the README.md to inform anyone
(somehow) finding this project about `npm update`.